chore: Update file paths in .gitignore
This commit updates the file paths in the .gitignore file to reflect the changes made in the server folder. The paths for the ignored files have been changed from 'server/myenv/' to 'python/myenv/', 'server/chroma/' to 'python/chroma/', and 'server/chroma.sqlite3' to 'python/chroma.sqlite3'. Additionally, the paths for the ignored files in the server folder have been changed to 'python/*.env', 'python/*.pyc', and 'python/__pycache__/'. The path for the ignored file 'client/src/test.jpg' remains unchanged, and the path for the ignored file 'server/main.py' has been changed to 'python/main.py'.